Account for arguments with only a type and no parameter name
Previously, the code was parsing the type into the argname variable and then
when it found something (not "," or ")") following the type it swapped the name
into the type and parsed next part as the name.

However, when no name is provided, as it allowed in c++, if you're not planning
on using the parameter, or it is in a function declaration, then the actual
type was being left in the name variable and the type variable was empty.

As a result function signatures for references were being generated without
knowledge of the type, which is the important factor in disambiguating
overloaded functions.

--HG--
extra : transplant_source : %5B%B7%F3%92%27%D3%B3%1C%E0%DBh%A2%3B%8A%83%E3%21%23b%95