Use term in description list as translation message
Here is a .rst that reproduces the problem fixed by this changes:

  term-word
    description-paragraph

"description-paragraph" is extracted as translation message by
"sphinx-build -b gettext" but "term-word" isn't.

It's because nodes.term node doesn't have source information.
Source information should be added by Docutils instead of Sphinx.

A patch that fix this problem in Doctuils had been submitted:
https://sourceforge.net/tracker/?func=detail&aid=3548418&group_id=38414&atid=422032

This changes are workarond to work with docutils-0.9.1 or earlier.

--HG--
branch : i18n-term-translatable