fix tests for autodoc novalue option
In the tests for autodoc the Options are of type struct,
while in the code there is a special autodoc.Options class,
which is a dict.

So "novalue" in self.options doesn't work,
but self.options.novalue does work for both.

Additionally the logic for autoattribute was wrong and is fixed now.

--HG--
branch : autodoc_novalue