Add +0000 to POT-Creation-Date
%z sets an empty string when datetimes are "naive" (no timezone
specified), but some packages expect a timezone specifier on the
POT-Creation-Date header and will error out if it is missing
(e.g. babel's pofile.read_po).

Lifted the stdlib's UTC timezone recipe and set this explicitly, it's
not the correct timezone but at least it's a timezone.

The LocalTimezone recipe might be a better one.

--HG--
branch : masklinn/add-0000-to-potcreationdate-z-sets-an-e-1369740575399