writers/latex: debug bibtex hypertarget labels
The LaTeX reference target used to be the bibitem id. But the bibitem id was
already rewritten to contain a number instead of the lowercased id name, that
is given by the bibtex file. This change adds a new element to the bibitem
instantiation which contains the id from the bibtex file. The hypertarget uses
this bibitem element now.

The alternative fix would be to hyperref the numbered id and leave the target
with the id number as it is. We decided against it, because in the html builder
both the ref and the target use the bibtex id name and not the number.

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>
Acked-by: Matthias C. M. Troffaes <matthias.troffaes@gmail.com>

--HG--
extra : source : 054d8e3d949ab13fdf2eb1a69a2c77d56f2c9370